+++ src/main/java/com/argentina_programa/Portfolio/controllers/ApiControllerSocial.java
@@ -9,6 +9,9 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import com.argentina_programa.Portfolio.services.interfaces.ISocialService;
+import io.jsonwebtoken.Jwts;
+import java.util.Base64;
+import javax.annotation.PostConstruct;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -16,6 +19,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestHeader;
 
 /**
  *
@@ -24,10 +28,14 @@ import org.springframework.web.bind.annotation.RequestBody;
 @RestController
 @RequestMapping("api/social")
 public class ApiControllerSocial {
-    
+    private String secret = "ArgProg";
     @Autowired 
     private ISocialService socialS;
     
+    @PostConstruct
+    protected void init (){
+        secret = Base64.getEncoder().encodeToString(secret.getBytes());
+    }
       
     @GetMapping(value = "/all", produces = MediaType.APPLICATION_JSON_VALUE)
     public ResponseEntity <Object> mostrarTodos(){
@@ -42,18 +50,35 @@ public class ApiControllerSocial {
         return ResponseEntity.ok(this.socialS.findByRedSocial(red_social));
     }
     @PostMapping ( value = "/save", produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity <Object> save (@RequestBody Social estudio){
-        this.socialS.save(estudio);
-        return ResponseEntity.ok(Boolean.TRUE);
+    public ResponseEntity <Object> save (@RequestBody Social estudio, @RequestHeader("Authorization") String token){
+        boolean bandera = getAuthentication(token);
+        if(bandera)
+            this.socialS.save(estudio);
+        return ResponseEntity.ok(bandera ? Boolean.TRUE: Boolean.FALSE);
     }
     @DeleteMapping ( value = "/{id}/delete")
-    public ResponseEntity <Object> delete ( @PathVariable int id){
-        this.socialS.deleteById(id);
-        return ResponseEntity.ok(Boolean.TRUE);
+    public ResponseEntity <Object> delete ( @PathVariable int id, @RequestHeader("Authorization") String token){
+        boolean bandera = getAuthentication(token);
+        if(bandera)
+            this.socialS.deleteById(id);
+        return ResponseEntity.ok(bandera ? Boolean.TRUE: Boolean.FALSE);
     }
     @PostMapping (value = "/{id}/update")
-    public ResponseEntity <Object> update ( @RequestBody Social estudio,@PathVariable int id){
-        this.socialS.update(estudio, id);
-        return ResponseEntity.ok(Boolean.TRUE);
+    public ResponseEntity <Object> update ( @RequestBody Social estudio,@PathVariable int id,@RequestHeader("Authorization") String token){
+        boolean bandera = getAuthentication(token);
+        if(bandera)
+            this.socialS.update(estudio, id);
+        return ResponseEntity.ok(bandera ? Boolean.TRUE: Boolean.FALSE);
+    }
+        private boolean getAuthentication (String token){
+        String user  ="";
+        if (token != null){
+            user = Jwts.parser()
+                    .setSigningKey(secret)
+                    .parseClaimsJws(token.replace("Bearer", "")) 
+                    .getBody()
+                    .getSubject();
+        }
+        return  user.equals("BrandonPrograma");
     }
 }
